requirements.txt
scikit-image
scikit-learn
scikit-metrics
scipy
numpy
torch
torchvision
tqdm
opencv-python
Pillow
kornia